pyproject.toml
[project]
name = "fastapi-demo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi-filter>=2.0.1",
    "fastapi-pagination[sqlalchemy]>=0.13.3",
    "fastapi[standard]>=0.116.1",
    "httptools>=0.6.4",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.41",
]
